services/worker/requirements.txt
# Celery and message broker
celery==5.3.4
redis==5.0.1

# Database (shared with API service)
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and security (for shared models)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client for external APIs
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-celery==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0